+++ socket/src/index.ts
@@ -1,14 +1,3 @@
-export {
-  Controller,
-  Connection,
-  Connect,
-  Disconnect,
-  GlobalEvent,
-  Event,
-  IO,
-  Socket,
-  Args,
-  Ack
-} from './decorators';
-export { IO_MIDDLEWARE, Middleware } from './middleware';
-export { attachControllers } from './socket';
+export * from './decorators';
+export * from './middleware';
+export * from './socket';